+++ src/components/cards/project-card.jsx
@@ -1,13 +1,14 @@
-import * as React from 'react';
-import { Box } from 'theme-ui';
+/** @jsx jsx */
+import { Box, Flex, jsx } from 'theme-ui';
 import SmallCard from './small-card';
-import { BadgeWithIcon } from 'components/badges/badge';
+import Badge, { BadgeWithIcon } from 'components/badges/badge';
 import ButtonWithIcon from 'components/buttons/button-with-icon';
+import LinkWithIcon from 'components/links/link-with-icon';
 
 const ProjectCard = ({
 	to = '/projects',
 	tag = 'Coding',
-	title = 'Skill Card',
+	title = 'Project Card',
 	children = 'Children',
 	borderColor = 'p700',
 }) => (
@@ -56,6 +57,80 @@ const ProjectCard = ({
 				Github
 			</ButtonWithIcon>
 		</Box>
+	</SmallCard>
+);
+
+export const LargeProjectCard = ({
+	title = 'Large Project Card',
+	github = 'https://github.com/',
+	preview = '/',
+	badges = ['Design', 'Programming', 'Web', 'Frontend'],
+	tag = 'Full Stack',
+	children = 'Children',
+	borderColor = 'p700',
+}) => (
+	<SmallCard borderColor={borderColor}>
+		<Box
+			as="header"
+			sx={{
+				fontSize: 24,
+				color: 'gs800',
+				fontFamily: 'head',
+				fontStyle: 'normal',
+				fontWeight: 'bold',
+				display: 'flex',
+				justifyContent: 'space-between',
+				alignItems: 'center',
+			}}
+		>
+			{title}
+			<BadgeWithIcon>{tag}</BadgeWithIcon>
+		</Box>
+		<Flex
+			sx={{
+				flexDirection: 'row',
+				my: '1.5rem',
+				gap: '5px',
+			}}
+		>
+			{badges.map((badgeText) => (
+				<Badge borderColor="warning" key={badgeText}>{badgeText}</Badge>
+			))}
+		</Flex>
+		<Box
+			sx={{
+				mb: '1.5rem',
+				fontFamily: 'body',
+				fontWeight: 'light',
+				fontSize: '0.875rem',
+				color: 'gs800',
+				textAlign: 'center',
+			}}
+			as="p"
+		>
+			{children}
+		</Box>
+		<Flex
+			sx={{
+				justifyContent: 'space-between',
+				alignItems: 'center',
+				fontFamily: 'body',
+				fontStyle: 'normal',
+				fontWeight: 'normal',
+				fontSize: '1rem',
+				p: 0,
+			}}
+			as="footer"
+		>
+			<ButtonWithIcon
+				icon="github"
+				side="left"
+				to={'https://github.com/' + github}
+			>
+				Github
+			</ButtonWithIcon>
+			<LinkWithIcon to={preview}>Preview</LinkWithIcon>
+		</Flex>
 	</SmallCard>
 );